+++ pkg/pd/pd.go
@@ -19,6 +19,7 @@ const (
 	DefaultUserAgent = "Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/79.0.3945.117 Safari/537.36"
 	// errors
 	ErrMissingPathToFile = "file path is required"
+	ErrMissingFileID     = "file id is required"
 )
 
 type ClientOptions struct {
@@ -79,6 +80,7 @@ func New(opt *ClientOptions, c *Client) *PixelDrainClient {
 }
 
 // UploadPOST POST /api/file
+// curl -X POST -i -H "Authorization: Basic <TOKEN>" -F "file=@cat.jpg" https://pixeldrain.com/api/file
 func (pd *PixelDrainClient) UploadPOST(r *RequestUpload) (*ResponseUpload, error) {
 	if r.PathToFile == "" {
 		return nil, errors.New(ErrMissingPathToFile)
@@ -104,7 +106,7 @@ func (pd *PixelDrainClient) UploadPOST(r *RequestUpload) (*ResponseUpload, error
 	}
 
 	// pixeldrain want an empty username and the APIKey as password
-	if r.Auth.APIKey != "" && !r.Anonymous {
+	if r.Auth.IsAuthAvailable() && !r.Anonymous {
 		addBasicAuthHeader(pd.Client.Header, "", r.Auth.APIKey)
 	}
 
@@ -127,6 +129,7 @@ func (pd *PixelDrainClient) UploadPOST(r *RequestUpload) (*ResponseUpload, error
 }
 
 // UploadPUT PUT /api/file/{name}
+// curl -X PUT -i -H "Authorization: Basic <TOKEN>" --upload-file cat.jpg https://pixeldrain.com/api/file/test_cat.jpg
 func (pd *PixelDrainClient) UploadPUT(r *RequestUpload) (*ResponseUpload, error) {
 	if r.PathToFile == "" {
 		return nil, errors.New(ErrMissingPathToFile)
@@ -141,16 +144,17 @@ func (pd *PixelDrainClient) UploadPUT(r *RequestUpload) (*ResponseUpload, error)
 		return nil, err
 	}
 
-	reqParams := req.Param{
-		"anonymous": r.Anonymous,
-	}
+	// we dont send this paramter due a bug of pixeldrain side
+	//reqParams := req.Param{
+	//	"anonymous": r.Anonymous,
+	//}
 
 	// pixeldrain want an empty username and the APIKey as password
-	if r.Auth.APIKey != "" && !r.Anonymous {
+	if r.Auth.IsAuthAvailable() && !r.Anonymous {
 		addBasicAuthHeader(pd.Client.Header, "", r.Auth.APIKey)
 	}
 
-	rsp, err := pd.Client.Request.Put(r.URL, pd.Client.Header, file, reqParams)
+	rsp, err := pd.Client.Request.Put(r.URL, pd.Client.Header, file)
 	if pd.Debug {
 		log.Println(rsp.Dump())
 	}
@@ -171,10 +175,54 @@ func (pd *PixelDrainClient) UploadPUT(r *RequestUpload) (*ResponseUpload, error)
 	return uploadRsp, nil
 }
 
-// GetFile GET /api/file/{id}
-func (pd *PixelDrainClient) GetFile() (*ResponseUpload, error) {
-	// todo
-	return nil, nil
+// Download GET /api/file/{id}
+func (pd *PixelDrainClient) Download(r *RequestDownload) (*ResponseDownload, error) {
+	if r.PathToSave == "" {
+		return nil, errors.New(ErrMissingPathToFile)
+	}
+
+	if r.ID == "" {
+		return nil, errors.New(ErrMissingFileID)
+	}
+
+	if r.URL == "" {
+		r.URL = fmt.Sprintf(APIURL+"/file/%s", r.ID)
+	}
+
+	// pixeldrain want an empty username and the APIKey as password
+	if r.Auth.IsAuthAvailable() {
+		addBasicAuthHeader(pd.Client.Header, "", r.Auth.APIKey)
+	}
+
+	rsp, err := pd.Client.Request.Get(r.URL, pd.Client.Header)
+	if pd.Debug {
+		log.Println(rsp.Dump())
+	}
+	if err != nil {
+		return nil, err
+	}
+
+	err = rsp.ToFile(r.PathToSave)
+	if err != nil {
+		return nil, err
+	}
+
+	fInfo, err := os.Stat(r.PathToSave)
+	if err != nil {
+		return nil, err
+	}
+
+	downloadRsp := &ResponseDownload{
+		StatusCode: rsp.Response().StatusCode,
+		FilePath:   r.PathToSave,
+		FileName:   fInfo.Name(),
+		FileSize:   fInfo.Size(),
+		ResponseDefault: ResponseDefault{
+			Success: true,
+		},
+	}
+
+	return downloadRsp, nil
 }
 
 // addBasicAuthHeader create a http basic auth header from username and password